apps/race-display/race_display_clean/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-Session==0.8.0
gunicorn==23.0.0
redis==5.0.7
psycopg2-binary==2.9.9
requests==2.32.3
beautifulsoup4==4.12.3
tinycss2==1.3.0
python-dotenv==1.0.1
Werkzeug==3.0.3
urllib3==2.2.2 